avoid exception if no group selected in adress book added error if attempt to compile in 64bits (exosip -and probably other parts- not 64bits clean)
git-svn-id: http://svn.assembla.com/svn/symphonie/trunk@145 57038eb5-193d-0410-9cfd-c93977fb03f4

+++ sipPhoneAbPlugin/sipPhoneAbPlugin.m
@@ -24,7 +24,7 @@
 	ABMultiValue* phones = [person valueForProperty:kABPhoneProperty];
 	int k=[phones count];
 	if (k<1) return nil;
-	unsigned int idx=[phones indexForIdentifier:identifier];
+	NSUInteger idx=[phones indexForIdentifier:identifier];
 	if ( NSNotFound ==idx) idx=0;
 	NSString *value=[phones valueAtIndex:idx];
 	return intl ? [value internationalCallNumber] : value ;
